+++ Backend/models/__init__.py
@@ -58,5 +58,6 @@ __all__ = [
     'TagCategory',
     'Tag',
     'TagLink',
-    'QuotesLast'
+    'QuotesLast',
+    'PasswordResetToken'
 ]